add: subcomponent of filterpanel
