+++ lib/components/FilterPanel/FilterPanelBottom.tsx
@@ -1,0 +1,9 @@
+const Bottom = () => {
+  return (
+    <div className="border border-t-0 rounded-bl-lg rounded-br-lg border-cu-black-100 px-4 py-4 sm:px-6 sm:py-5">
+      Bottom
+    </div>
+  )
+}
+
+export default Bottom
